LPS-55833 Moves java
